# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GlossoGen Simulation Server",
    "version": "0.1.0"
  },
  "paths": {
    "/api/g/{group_slug}/runs": {
      "get": {
        "summary": "List Runs",
        "description": "List one page of simulation runs owned by the active group, newest-first.\n\nFilters: ``scenario`` keeps runs in any of the listed scenarios (OR\nsemantics); ``labels`` keeps runs carrying every listed label (AND\nsemantics); ``run_id_contains`` keeps runs whose ``scenario/run_dir_name``\nid contains the substring (case-insensitive); ``status`` restricts to a\nfinal status; ``contains_agent_id`` keeps runs that registered that agent\n(used by the cross-run replace-agent picker). ``offset``/``limit`` page the\nresult; ``total`` is the count matching the filters before paging.",
        "operationId": "list_runs_api_g__group_slug__runs_get",
        "parameters": [
          {
            "name": "scenario",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "title": "Scenario"
            }
          },
          {
            "name": "contains_agent_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Contains Agent Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/RunStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "labels",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "title": "Labels"
            }
          },
          {
            "name": "run_id_contains",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Run Id Contains"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/g/{group_slug}/runs/{scenario}/{run_dir_name}": {
      "get": {
        "summary": "Get Run Detail",
        "description": "Get full detail for a specific simulation run.",
        "operationId": "get_run_detail_api_g__group_slug__runs__scenario___run_dir_name__get",
        "parameters": [
          {
            "name": "scenario",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scenario"
            }
          },
          {
            "name": "run_dir_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Dir Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Run",
        "description": "Stop the simulation if still running, then move the run directory to trash.\n\nReversible: the run's files are moved into ``{runs_dir}/_trash/`` rather\nthan removed, so a deletion can be undone by moving the directory back and\nre-registering it. The Postgres index row is deleted so the run no longer\nappears in listings.",
        "operationId": "delete_run_api_g__group_slug__runs__scenario___run_dir_name__delete",
        "parameters": [
          {
            "name": "scenario",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scenario"
            }
          },
          {
            "name": "run_dir_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Dir Name"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/g/{group_slug}/runs/{scenario}/{run_dir_name}/export/bundle": {
      "get": {
        "summary": "Export Run Bundle",
        "description": "Export a simulation run as a tar.gz bundle.",
        "operationId": "export_run_bundle_api_g__group_slug__runs__scenario___run_dir_name__export_bundle_get",
        "parameters": [
          {
            "name": "scenario",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scenario"
            }
          },
          {
            "name": "run_dir_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Dir Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tar.gz bundle of the simulation run.",
            "content": {
              "application/json": {
                "schema": {}
              },
              "application/gzip": {}
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/g/{group_slug}/runs/{scenario}/{run_dir_name}/export/zip": {
      "get": {
        "summary": "Export Run Zip",
        "description": "Export a simulation run as a ``{run_dir_name}.zip`` for manual extraction.",
        "operationId": "export_run_zip_api_g__group_slug__runs__scenario___run_dir_name__export_zip_get",
        "parameters": [
          {
            "name": "scenario",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scenario"
            }
          },
          {
            "name": "run_dir_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Dir Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Zip archive of the simulation run, nested under a run-id folder.",
            "content": {
              "application/json": {
                "schema": {}
              },
              "application/zip": {}
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/g/{group_slug}/runs/import": {
      "post": {
        "summary": "Import Run Bundle",
        "description": "Import a simulation run from an exported bundle tar.gz.\n\nIdempotent: if a run with the same run_id already exists, returns the\nexisting run without re-importing.",
        "operationId": "import_run_bundle_api_g__group_slug__runs_import_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_import_run_bundle_api_g__group_slug__runs_import_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportBundleResponse"
                }
              }
            }
          },
          "422": {
            "description": "Invalid or incomplete bundle."
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_import_run_bundle_api_g__group_slug__runs_import_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_import_run_bundle_api_g__group_slug__runs_import_post"
      },
...
      "ForkSource": {
        "properties": {
          "source_run_id": {
            "type": "string",
            "title": "Source Run Id"
          },
          "target_message_id": {
            "type": "string",
            "title": "Target Message Id"
          },
          "forked_at": {
            "type": "string",
            "format": "date-time",
            "title": "Forked At"
          }
        },
        "type": "object",
        "required": [
          "source_run_id",
          "target_message_id",
          "forked_at"
        ],
        "title": "ForkSource",
        "description": "Provenance information for a forked simulation run."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ImportBundleResponse": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "scenario_name": {
            "type": "string",
            "title": "Scenario Name"
          },
          "run_dir": {
            "type": "string",
            "title": "Run Dir"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "scenario_name",
          "run_dir"
        ],
        "title": "ImportBundleResponse",
        "description": "Response after successfully importing a run bundle."
      },
...
      "RunListResponse": {
        "properties": {
          "runs": {
            "items": {
              "$ref": "#/components/schemas/RunSummary"
            },
            "type": "array",
            "title": "Runs"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "runs",
          "total"
        ],
        "title": "RunListResponse",
        "description": "Response model for the paginated runs endpoint.\n\n``runs`` is one page (newest-first); ``total`` is the number of runs\nmatching the request's filters before paging, for the load-more UI."
      },
      "RunStatus": {
        "type": "string",
        "enum": [
          "scenario_complete",
          "in_progress",
          "starting",
          "error",
          "killed"
        ],
        "title": "RunStatus",
        "description": "Status of a simulation run."
      },
      "RunSummary": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "scenario_name": {
            "type": "string",
            "title": "Scenario Name"
          },
          "scenario_description": {
            "type": "string",
            "title": "Scenario Description"
          },
          "scenario_config": {
            "additionalProperties": true,
            "type": "object",
            "title": "Scenario Config"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "total_messages": {
            "type": "integer",
            "title": "Total Messages"
          },
          "total_cost_usd": {
            "type": "number",
            "title": "Total Cost Usd"
          },
          "duration_seconds": {
            "type": "number",
            "title": "Duration Seconds"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus"
          },
          "has_evaluation": {
            "type": "boolean",
            "title": "Has Evaluation"
          },
          "evaluation_in_progress": {
            "type": "boolean",
            "title": "Evaluation In Progress"
          },
          "run_dir": {
            "type": "string",
            "title": "Run Dir"
          },
          "fork_source": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ForkSource"
              },
              {
                "type": "null"
              }
            ]
          },
          "replace_agent_source": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReplaceAgentSource"
              },
              {
                "type": "null"
              }
            ]
          },
          "cross_run_replace_agent_source": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CrossRunReplaceAgentSource"
              },
              {
                "type": "null"
              }
            ]
          },
          "resume_at_round_source": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResumeAtRoundSource"
              },
              {
                "type": "null"
              }
            ]
          },
          "models": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Models"
          },
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "agent_models": {
            "items": {
              "$ref": "#/components/schemas/AgentModelSummary"
            },
            "type": "array",
            "title": "Agent Models"
          },
          "labels": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Labels"
          },
          "has_note": {
            "type": "boolean",
            "title": "Has Note"
          },
          "current_round": {
            "type": "integer",
            "title": "Current Round"
          },
          "evaluation_content_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evaluation Content Hash"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "scenario_name",
          "scenario_description",
          "scenario_config",
          "timestamp",
          "total_messages",
          "total_cost_usd",
          "duration_seconds",
          "status",
          "has_evaluation",
          "evaluation_in_progress",
          "run_dir",
          "fork_source",
          "replace_agent_source",
          "cross_run_replace_agent_source",
          "resume_at_round_source",
          "models",
          "provider",
          "agent_models",
          "labels",
          "has_note",
          "current_round",
          "evaluation_content_hash"
        ],
        "title": "RunSummary",
        "description": "Summary of a single simulation run for the runs list endpoint."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}